02-cloud-native-tooling/python/summarize-yt-ai/requirements.txt
yt-dlp
openai-whisper
ffmpeg
langchain